final project

+++ frontend/src/components/ShoppingItem.jsx
@@ -10,21 +10,25 @@ function ShoppingItem(props) {
 
     const registerToBuy = () => {
         if (props.loggedIn.currentLoggedIn !== props.item.itemOwner) {
-        const options = {
-            method: 'GET',
-            url: `http://127.0.0.1:8000/registerToBuy/${props.loggedIn.currentLoggedIn}/${props.item.itemId}/${bidAmount}`,
-            headers: { accept: 'application/json' }
-        }
-
-        axios
-            .request(options)
-            .then(response => {
-                props.setResponse([response.data.result, "success"])
-                props.deleteItem()
-            })
-            .catch(error => {
-                console.log(error)
-            });
+            if (bidAmount > props.item.itemPrice) {
+                const options = {
+                    method: 'GET',
+                    url: `http://127.0.0.1:8000/registerToBuy/${props.loggedIn.currentLoggedIn}/${props.item.itemId}/${bidAmount}`,
+                    headers: { accept: 'application/json' }
+                }
+    
+                axios
+                    .request(options)
+                    .then(response => {
+                        props.setResponse([response.data.result, "success"])
+                        props.deleteItem()
+                    })
+                    .catch(error => {
+                        console.log(error)
+                    });
+            } else {
+                props.setResponse(["You cannot bid lower than the current price", "warning"])
+            }
         } else {
             props.setResponse(["You cannot buy your own item", "danger"])
             props.deleteItem()
@@ -50,7 +54,7 @@ function ShoppingItem(props) {
             </td>
             {/* item price */}
             <td>
-                <div>{props.price}</div>
+                <div>{props.item.itemPrice}</div>
             </td>
             {/* bidding item */}
             <td>
